python-backend/requirements.txt
# Quantalyze Python Backend - Requirements
# FastAPI Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.3.0
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# AI/ML & Data Processing
openai==1.10.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# PDF & Report Generation
reportlab==4.0.8
openpyxl==3.1.2
python-docx==1.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
jinja2==3.1.3
Pillow==10.2.0

# CORS
starlette==0.35.1
